refactor: FollowList 컴포넌트에서 UserList 컴포넌트 분리(#27)
