+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './PostCard/PostCard';
 export * from './RecommendFollow/RecommendFollow';
 export * from './RecommendFollow/FollowList';
+export * from './RecommendFollow/UserList';
